scroll to replied message, scroll to bottom, split messages to smaller components
