+++ package.json
@@ -41,6 +41,7 @@
     "lodash": "^4.17.21",
     "moment": "^2.30.1",
     "moment-timezone": "^0.5.45",
+    "motion": "^12.4.7",
     "rc-progress": "^4.0.0",
     "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
